refactor: move test utils to package
