+++ tests/parallels/contact-us.spec.ts
@@ -1,8 +1,7 @@
 import { test } from '@affine-test/kit/playwright';
+import { openHomePage } from '@affine-test/kit/utils/load-page';
+import { waitEditorLoad } from '@affine-test/kit/utils/page-logic';
 import { expect } from '@playwright/test';
-
-import { openHomePage } from '../libs/load-page';
-import { waitEditorLoad } from '../libs/page-logic';
 
 test('Click right-bottom corner contact icon', async ({ page }) => {
   await openHomePage(page);
