+++ tests/parallels/image-preview.spec.ts
@@ -1,13 +1,12 @@
-import type { Page } from '@playwright/test';
-import { expect, test } from '@playwright/test';
-import fs from 'fs';
-
-import { openHomePage } from '../libs/load-page';
+import { openHomePage } from '@affine-test/kit/utils/load-page';
 import {
   getBlockSuiteEditorTitle,
   newPage,
   waitEditorLoad,
-} from '../libs/page-logic';
+} from '@affine-test/kit/utils/page-logic';
+import type { Page } from '@playwright/test';
+import { expect, test } from '@playwright/test';
+import fs from 'fs';
 
 async function importImage(page: Page, url: string) {
   await page.evaluate(
