+++ tests/parallels/local-first-avatar.spec.ts
@@ -1,9 +1,8 @@
 import { test } from '@affine-test/kit/playwright';
+import { openHomePage } from '@affine-test/kit/utils/load-page';
+import { newPage, waitEditorLoad } from '@affine-test/kit/utils/page-logic';
+import { assertCurrentWorkspaceFlavour } from '@affine-test/kit/utils/workspace';
 import { expect } from '@playwright/test';
-
-import { openHomePage } from '../libs/load-page';
-import { newPage, waitEditorLoad } from '../libs/page-logic';
-import { assertCurrentWorkspaceFlavour } from '../libs/workspace';
 
 test('should create a page with a local first avatar', async ({ page }) => {
   await openHomePage(page);
