+++ tests/parallels/local-first-favorite-page.spec.ts
@@ -1,15 +1,14 @@
 import { test } from '@affine-test/kit/playwright';
-import { expect } from '@playwright/test';
-
-import { openHomePage } from '../libs/load-page';
+import { openHomePage } from '@affine-test/kit/utils/load-page';
 import {
   clickPageMoreActions,
   getBlockSuiteEditorTitle,
   newPage,
   waitEditorLoad,
-} from '../libs/page-logic';
-import { waitForLogMessage } from '../libs/utils';
-import { assertCurrentWorkspaceFlavour } from '../libs/workspace';
+} from '@affine-test/kit/utils/page-logic';
+import { waitForLogMessage } from '@affine-test/kit/utils/utils';
+import { assertCurrentWorkspaceFlavour } from '@affine-test/kit/utils/workspace';
+import { expect } from '@playwright/test';
 
 test('New a page and open it ,then favorite it', async ({ page }) => {
   await openHomePage(page);
