+++ tests/parallels/local-first-trash-page.spec.ts
@@ -1,13 +1,12 @@
 import { test } from '@affine-test/kit/playwright';
-import { expect } from '@playwright/test';
-
-import { openHomePage } from '../libs/load-page';
+import { openHomePage } from '@affine-test/kit/utils/load-page';
 import {
   getBlockSuiteEditorTitle,
   newPage,
   waitEditorLoad,
-} from '../libs/page-logic';
-import { assertCurrentWorkspaceFlavour } from '../libs/workspace';
+} from '@affine-test/kit/utils/page-logic';
+import { assertCurrentWorkspaceFlavour } from '@affine-test/kit/utils/workspace';
+import { expect } from '@playwright/test';
 
 test('New a page , then delete it in all pages, finally find it in trash', async ({
   page,
