+++ tests/parallels/subpage.spec.ts
@@ -1,8 +1,7 @@
 import { test } from '@affine-test/kit/playwright';
+import { openHomePage } from '@affine-test/kit/utils/load-page';
+import { waitEditorLoad } from '@affine-test/kit/utils/page-logic';
 import { expect } from '@playwright/test';
-
-import { openHomePage } from '../libs/load-page';
-import { waitEditorLoad } from '../libs/page-logic';
 
 test('Create subpage', async ({ page }) => {
   await openHomePage(page);
