+++ tests/parallels/theme.spec.ts
@@ -1,10 +1,9 @@
 import { resolve } from 'node:path';
 
 import { test, testResultDir } from '@affine-test/kit/playwright';
+import { openHomePage } from '@affine-test/kit/utils/load-page';
+import { waitEditorLoad } from '@affine-test/kit/utils/page-logic';
 import { expect } from '@playwright/test';
-
-import { openHomePage } from '../libs/load-page';
-import { waitEditorLoad } from '../libs/page-logic';
 
 // default could be anything, according to the system
 test('default white', async ({ browser }) => {
